test(ChatMessageRepositoryTest): migrate from extention to abstraction based configure
